fixed issues related to guardian ship coming back after being defeated.. 1. when the ship is plundered, it is now removed from the Orion system so that hasMonster() will return false. In addition, all empires get a system refresh so that they know it's no longer guarded and can react appropriately
2. the obsolete "isGuarded()" method on StarSystem was removed and
references to it were replaced with "hasMonster()" with the exception of
the Orion council bonus calculation, which checks for an owned system
being Orion as opposed to guarded

3. the "endofCombat(true)" call was placed in the "combatIsFinshed()"
method because it always needs to be called when that method returns
true. There was call for enabling a UI button that was not triggering
endOfCombat(true) which prevented the Guardian ship from being plundered
which is where it is removed form the system

4. I never write comments this detailed but shit this is a public repo
now so I have to pretend to be a diligent programmer.