fun one. AI ships with just missile weapons were not retreating whne they were out of ammo. This was because the estimatedKills method was not considering limited shot weapons (missiles) running out. Fixed that.
But the save game also showed that a pre-beta change to "spread out"
weapon counts on AI ship designs (i.e 5 weapons in 4 slots instead of 20
in 1 slot) was breaking the weapons.getIndexOf(component) calls since all of
the components were the same! This prevented weapon cycling and resulted
in AI ships only firing one of the weapon slots instead of all four!

Yeah, players have had it lucky for 48 hours, but no more.

Instead of keeping a "selectedWeapon" var (ShipComponet) on each combat
stack to designate the currently used, I changed it to now use a
"selectedWeaponIndex" var (int) so that the indexOf() lookups were
unnecessary. Now the AI ships cycle through all of their weapons
properly and fire them all. There is still an issue where multiple
missile stacks are on top of each other on the combat UI, but that is
a display issue and can be fixed with a later change.

Note: the rotateToUseableWeapon method has apparently never ever worked
but the issue never came up because AIShipDesigner previously never
spread out the same weapons across 4 banks. That change (spreading out)
was made to take advantage of recent AIShipCaptain changes of allowing
AI ships to be able to attack a 2nd target if they have killed another
target on the same turn.

Hope you enjoyed the description. I'm trying to be good here.