multiple changes to improve feel of fleets and their selection on the map: 1. The map no longer sends double-clicks to the selection panel. None of them use this anymore and the extra click events sent to the panels would trigger behavior unintuitive to the user.
2. AI Fleets can no longer be selected, just hovered. (like transports)

3. When a player fleet is selected, all system hovering events are
handled. Previously those systems where the fleet could not be sent (out
of range, etc), the event would be kicked back and would trigger a
display of the system which created a jarring flashing back and forth if
the fleet you had selected was within the hit box of a system. Now these
systems are handled by the fleet panel with an appropriate error (out of
range, etc).

4. Attempting to send a retreating fleet to an invalid destination
(non-allied system or not on the turn it was retreating) will show an
explicit "Invalid Retreat" message instead of the "Cancel" button

5. When hovering over fleets, the map no longer returns the first ship
in the list of visible ship whose hit box contains the mouse location.
Now it returns the ship whose hit box is most closely centered with the
mouse location. This makes it much easier to select particular fleets in
a crowded area.