related to previous commit
1. When a player fleet is selected, never check for hovering over
fleets. This is to make it easy for players to send fleets to other
systems that may be crowed with fleets. Even if the fleet cannot be sent
anywhere, this behavior will trigger the correct error message when
hovering over systems.

To unselect a fleet, right-click or hit the cancel button.