Performance Improvement while drawing crusters (borders)
Instead of making a call to graphics2D to draw each border and external border of each system the player owns (and respective allies), this implementation uses the concept of Areas. All borders are first added to the area and only then the resulting border is drawn with graphics2D.

This is considerable faster and can be noticed more and more as your game progresses and the number of systems you own increses.