New map shape, Swirl Clusters
Swirl Clusters is a map shape with a single thin spiral from the center. There are numerous clusters located along the spiral.

The length of the spiral, number of spiral twists, number of clusters, and cluster radius all scale off of the map size.

Map generation time can be very long (due to upwards of a few thousand trig function calls for every star generated). 10-15 seconds for the 100,000 star map on my system.