New map shape, Grid
Grid is a map shape with a rectangular grid of stars and random clusters of stars at some intersections.

The number of grid lines, number of clusters, and cluster radius all scale off of the map size.

Map generation time can be very long (due to how stars are selected to be viable). 10-15 seconds for the 100,000 star map on my system.