Flight path line and Rally path line changes
Slightly increased the thickness of fleet flight path and rally path. Slight change in the line dash distances to work with the slightly thicker lines.

The rally path animation was fixed with a change in the line dash distances. Because while rallyStroke gets defined with 9 "frames", the actual animation call is only mod%6.