Xilmi ai (#45)
* Replaced many isPlayer() by isPlayerControlled()

This leads to a more fluent experience on AutoPlay.
Note: the council-voting acts weird when you try that there, getting you stuck, and the News-Broadcasts also seem to use a differnt logic and are always shown.

* Orion-guard-handling

Avoid sending colonization-fleets, when guardian is still alive
Increase guardian-power-estimate to 100,000
Don't send and count bombers towards guardian-defeat-force

* Fixed Crash from a report on reddit

Not sure how that could happen though. Never had this issue in games started on my computer.

* Fixed crash after AI defeats orion-guard

Tried to access an out-of-bounds-value because it couldn't find the System where the guardian was. So looking to find the Planet with the Orionartifact now, which does find it.

* All the forgotten isPlayer()-calls

Including the Base- and Modnar-AI-Diplomat

* Personalities now play a role in my AI once again

Instead of everyone having the same conditions for war, there's now different ones. Some are more aggressive, some are less.

* Fixed issue where enemy-fleets weren's seen, when they actually were.

Also added a way of guessing when enemy-fleets aren't seen so it shouldn't trickle in tiny fleets that all retreat anymore in the early-game, before they can see the enemy-fleets. And also not later because they now actually can see them.

* Update AIDiplomat.java

* Delete ShipCombatResults.java

* Delete Empire.java

* Delete TradeRoute.java

* Revert "Delete TradeRoute.java"

This reverts commit 6e5c5a8604e89bb11a9a6dc63acd5b3f27194c83.

* Revert "Delete Empire.java"

This reverts commit 1a020295e05ebc735dae761f426742eb7bdc8d35.

* Revert "Delete ShipCombatResults.java"

This reverts commit 5f9287289feb666adf1aa809524973c54fbf0cd5.

* Update ShipCombatResults.java

reverting pull-request... manually because I don't know of a better way 8[

* Update Empire.java

* Update TradeRoute.java

* Merge with Master

* Update TradeRoute.java

* Update TradeRoute.java

github editor does not allow me to remove a newline oO

* AI improvements

Added parameter to make the colonizer-function return just the uncolonized planets without substracting the colony-ships.

Impelemented a very differnt diplomatic behavior, that supposedly is more suitable to actually winning by waging less war, particularly when big enough to be voted for.

Revoked that AI doesn't adjust to enemy-troops when it has a more defensive unit-composition. In lategame going by destructive-power of bombers hasn't much merit, when 1 bombers can pack 60ish neutronium-bombs.

Fixed an issue that prevented designing extended-fuel-range-colonizers asap. This once again helps expansion-speed a lot.

Security now is only spent against empires actually in range to spy, as it was intended. Contact via council should not increase paranoia. Also Xenophobic-extra-paranoia removed.

* Update OrionGuardianShip.java

Co-authored-by: rayfowler <58891984+rayfowler@users.noreply.github.com>