AI and damage-simulation-stuff
No longer bombing colonies of someone who isn't my enemy
Asking others to join a war, when our opponent is stronger than us
Agreeing to join-war request when not busy otherwise, together we'd be stronger and the asked-for-faction is our primary target anyways
Restricted war over espionage even more, to only do it if we feel ready. Much better to share a few techs than to die.
Wars for strategic reasons can now once again be cancelled.
Doing a cost:gain-analysis for invasions, that is much more likely to invade, especially if we have previously been invaded and the planet has factories from us. It also takes the savings of not having to build a colony-ship into account.
Further reduced the overkill for invasions. We want to do more of them when they are cost-efficient but still not waste our population.
More planets will participate in the construction of colony-ships.
When we are at war or have others in our range during the phase where we want to churn out colony-ships, we mix in combat-ships to make us less of a juicy target and be able to repel sneak-attacks better.
Hostile-planets now get more population sent to.
Stacks of strong individual ships about to die in combat, will consider their lessened chances of survival and may now retreat when damaged too much.
Reverted change in method that I don't use anyways.
Fixed shield-halving-properties of weapons like neutron-pellet-guns not being taken into account in combat-simulations using the firepower-method.
Fixed bombers estimating the whole stack sharing a single bomb, which dramatically underestimated the firepower they had.