AI performance
Instead of iterating through a triple-nestes-loop of fleets, systems and fleets, the inner loop is now buffered so that in consecutive calls there's a lot less to process. This significantly improves turn-times.
Colonizers with extended range can now be sent to colonize something when they are in a fleet that otherwise couldn't travel so far.
Removed unused code from AIGeneral.
AIShipCaptain is now thread-save due to no longer manipulating stack and instead strong the target in their own local member-variable.
Specials are now properly used in the right order. All specials except repulsor-beam and stasis-field are now cast before the other weapons.
Now fleeing from ship with repulsor if our optimal-range is outranged, not just our maximum-range. Also now considering that the repulsor-stack can protect more than itself (up to 3 more stacks when they are in a corner).
Will now design and use huge colonizers when stranded in a constellation that otherwise can't be escaped. The logic to determine that might still be a little flawed, though.
Repulsor now will be countered even if it is only researched and not installed in a ship yet since only reacting once it's seen could be too late.
Made shipComonentCanAttack public for improved logic in attacking.
Included fix from /u/albertinix that caused a crash when hitting pause in auto-combat.
UncolonizedPlanetsInRange-Methods in empire.java now actually only return uncolonized planets.
Fixed an issue where shield-reduction would be applied twice in combat-simulation.
Capped the amount of estimated kills at the amount of targets destroyed.