Blackhole-fixes
Fixed an issue that stacks killed by Blackhole-Generator only die in combat but are still alive afterwards.
Fixed an issue that Blackhole-Generator would always kill at least one ship, even if it rolled low. It now can miss when the amount of enemy ships is smaller than the percentage it rolls to kill.
Fixed how damage of blackhole-generator was shown as the percentage of the hitpoints of the ships rather than the total hitpoints of all ships it killed.