AI improvements
No longer breaking trade-treaty with war-target to avoid giving a warning that could be used to prepare.
Fixed crash related to sorting a list with random values that could change during the lifetime of the list. That random wasn't overly useful anyways.
Autoresolve will no longer ignore repulsors (but only for Xilmi-AI so far).
Ships will now always try to come as close as they can before shooting as a means to increase hit-chance, use short-range-specials or make it harder to dodge missiles.
Added back the value of the participants for retreat-decisions, because otherwise a clearly won battle could be given up because it was slighlty cost-inefficient.
No longer always countering repulsors just because someone could use them. Instead now looking at whether they are actually in use before the decision is made.