ensure information about scheduled transports is saved, even if the transport hasn't launched yet.
This required moving three values from the non-serialized
SystemTransportSprite class to the StarSystem class, and then ensuring
they are synched properly during gameplay, then properly reset to the
SystemTransportSprite when it is lazy-initted after reloading.

These are the kind of non-optimal changes you have to make late in a
program's development cycle in order to avoid rewriting a lot of code or
breaking existing saves.