on the fleet screen, the mass transport dialog now has a new flag column for each system you are transport from.
On this dialog and now on the colony listing, clicking the flag column
will update the flag color for the system. Previously, this column had
been display-only.

Finally, right-clicking on flags now (anywhere) will reset them back to
"no flag" instead of going reverse order. This is because there are many
flag colors now so an easy reset is needed.