when alliances and pacts are broken because of spies being caught, a new label [Your spies were caught.] is now added to the incident description.
This is also appended to the OathBreak notice the player receives when
this occurs, so he will better understand why the alliance or pact is
being dissolved.

Finally, observing races who notice the oathbreaking will no longer
notify the player that they, too, don't like oathbreaking. However, the
oath break will still be in their diplomatic incident as expected.