add a "Combat Stats: ON/OFF" button on the combat UI to make it clear that the tactical view can be toggled off and on. Space Monsters will show their health bar when this is enabled.
For foreign languages, I used the "Combat Stats" title on the Design UI
for the name, and then either ON/OFF or yes/no for the right side of the
colon. My apologies to any translators if I did this incorrectly! I was
trying to use previously translated lines from other parts of the game