corrected a bug where planet could exceed population limits when captured by an excessive number of incoming enemy transports. This check was already being made when friendly transports joined a colony. Thanks ajkfreelance for bringing this to my attention!
Also, removed the capturedOrion() method in Colony.java along with the
orionEmpire() methods in Galaxy.java. These methods are never
used/called as the Guardian was ultimately implemented as a Space Monster.

The orionEmpire instvar on Galaxy.java, while unused, needs to remain on
the class to preserve save-game compatibility