a couple fixes related to abandoning colonies. First, the max transports allowed is now properly capped in case the player had already abandoned the colony. That allows the player to exceed the normal transport limit so the tranport pane would correct for this if an abandon request was immediately follow up by a transport request.
Secondly, receiving transports are now properly capped by the receiving
planet size. This is similar to a fix recently made for invading
transports. This cap is done by the calling method rather than being
encapsulated in the receiving method. Many different functions call the
receiving method and I do not want it to automatically correct bad input. I want bad input to
cause an error so that it can be reported, properly tracked down, and fixed.