reset the spyTime for SystemView when the empire of a system changes.
This corrects a problem where you had spy information for an alien
system (i.e. pop = 15) and the system is subsequently captured by
another race. The spyTime variable should be reset, indicating that not
only is your spy info out of date, but it is irrelevant. It should now
show pop = "???" until you have spy networks in the second race.