Remove ship range fill
Colouring the entire range with empire colour means that a lot of detail of the map is lost - most importantly this makes it really hard to see where the nebulas begin for certain empire colours. Removing the fill still leaves the outline retaining the information but removes some of the noise created by the fill.