Make it easier to select star system
I have generally found that the selection bounds for fleets is too large and selection bounds for star system is too small - especially when dealing with star systems with a lot of traffic in them. i.e. I constantly select fleets when I just want to click on the system. This change adjusts both to improve this experience.