Show Help only if requested
Having Help UI show up at the start of every game adds needless clicks to the majority of games.

While it could be better to only show the help UI for the first game ever played the Help UI is probably better replaced with tool tips anyway so no work has been done on this here.