Stop star highlights overlapping
Further adjustment to selection circles and background fill for star systems to stop them from overlapping. This change doesn't guarantee that overlaps are impossible as star system placement is not yet understood but does seem to make it less likely.

Moved some variables to a common location to avoid having to adjust as many variables to achieve a resizing of the fill/selection circles.