Reduce log spam during start up
Because the licencing for some clear text files is not defined I have opted to not change them (with the goal of replacing them completely with explicitly licenced ones in the future). Due to some other changes made some values in these clear text files are no longer used and this gets logged to console. Since the log lines are not helpful and I can't change the underling file - remove the line responsible for the logging.