Remove relationship masking
Knowing the relationship between you and another empire is crucial to
being able to engage with the diplomacy part of the game (in my
opinion). Therefore, the omission of this information when playing
against the Darloks serves to only worsen the experience (again in my
opinion). (As this is my fork my opinion is the deciding factor.)