Remove Final War
One of genius things about the original game is that it realised that
cleaning up the map after you had already won is really boring. So the
victory condition is outnumbering your opponents 2:1 (either by
yourself or as a coalition) - thus removing the later boring third of
the game.

Allowing AI players to rebel against the council fundamentally
undermines this. As such this functionality has been completely
removed.

This change goes a step further by removing the players ability to
rebel against the council. Rebelling against the council put you back
in the degenerate state of having to play until all of the map is
painted one colour which is boring. While I understand that this gives
the player an opportunity for "revenge" which probably help to make
the experience of losing more palatable - I believe that leaving the
sting of defeat un-balmed (as it were) helps to make the experience
more memorable, even if unpleasant. It also simplifies the code which
is an aim of this fork.