Remove memory dump option
The behaviour of this option doesn't seem consistent/is buggy.
Instead of trying fix it, simply remove it.

The logging of these kind of metrics is probably best handled in a
completely different way (dev flags, logging, etc) so, if there is a
need for it, it is unlikely that I would like to resurrect this
feature as it is currently implemented.