Remove canSendTransports() check
This check is incorrect as it is perfectly valid to send an invasion
transports while you have only one colony.

More broadly, since this guards access to further menus (not the
action itself) it is better removed so that the menu can be explored
even when requirements to use it are not met - allowing the user to
experiment and discover features they might otherwise miss.