Rework default system name overlay
I haven't found knowing system population and factories useful at a
glance and at the size it was rendered it was hard to read. Therefore,
remove population and factories display.

This leaves the system name display looking a bit naff so it has been
restyled. It may need further updates but this is good enough for now.