Rework Zoom In and Out Controls
I think the background colour of both of those buttons must be a
mistake as it looks bad and doesn't fit in with the rest of the
Controls. Therefore, it has been changed to black (like the other
Controls).

The Zoom Controls have also been moved to the top of the Control
"bit". I struggle to explain why but this seems more intuitively the
correct place for them.