Remove Paranoia Incident
Making the relations worse based on how many systems have been taken
from an empire encourages you to fight an empire to the death (or as
close to it as possible) since you will never have as good relations
with them as before. Fighting an empire to the death (or until their
threat is trivial) makes for a less interesting diplomatic game
because the number of players is effectively reduced making it
simpler and less engaging. To that end remove Paranoia Incident.

Also, it's not paranoia if they are out to get you.